di github

+++ app/src/main/java/com/example/comffee/Item.kt
@@ -3,6 +3,6 @@ package com.example.comffee
 data class Item(
     var item_id : String ?= null,
     var nama_barang : String ?= null,
-    var harga : Long ?= null,
+    var harga : Double ?= null,
     var imagePath : String ?= null
 )
